+++ vignettes/link2GI2.R
@@ -5,7 +5,6 @@
 #  saga
 
 ## ----eval=FALSE---------------------------------------------------------------
-#  # find all SAGA GIS installations at the default search location
 #  require(link2GI)
 #  grass <- link2GI::findGRASS()
 #  grass
@@ -13,17 +12,16 @@
 #  otb
 
 ## ----eval=FALSE---------------------------------------------------------------
-#  # find all SAGA GIS installations at the default search location
 #  require(link2GI)
-#  link2GI::initProj(projRootDir = tempdir(),
-#                   projFolders = c("data/",
-#                                   "data/level0/",
-#                                   "data/level1/",
-#                                    "output/",
-#                                    "run/",
-#                                    "fun/"),
-#                   path_prefix = "path_to_" ,
-#                   global =TRUE)
+#  envrmt = link2GI::createFolders(root_folder = tempdir(),
+#                                  folders = c("data/",
+#                                              "data/level0/",
+#                                              "data/level1/",
+#                                              "output/",
+#                                              "run/",
+#                                              "fun/"),
+#                                  path_prefix = "path")
+#  envrmt
 
 ## ----eval=FALSE, echo=FALSE, message=FALSE, warning=FALSE---------------------
 #  # find all SAGA GIS installations at the default search location
@@ -72,18 +70,18 @@
 #   require(sf)
 #  
 #   # proj folders
-#   projRootDir<-tempdir()
-#   paths<-link2GI::initProj(projRootDir = projRootDir,
-#                            projFolders = c("project1/"))
+#   root_folder<-tempdir()
+#   paths<-link2GI::createFolders(root_folder = root_folder,
+#                            folders = c("project1/"))
 #  
 #   # get  data
 #   nc <- st_read(system.file("shape/nc.shp", package="sf"))
 #  
-#   # CREATE and link to a permanent GRASS folder at "projRootDir", location named "project1"
-#   linkGRASS(nc, gisdbase = projRootDir, location = "project1")
+#   # CREATE and link to a permanent GRASS folder at "root_folder", location named "project1"
+#   linkGRASS(nc, gisdbase = root_folder, location = "project1")
 #  
-#   # ONLY LINK to a permanent GRASS folder at "projRootDir", location named "project1"
-#   linkGRASS(gisdbase = projRootDir, location = "project1", gisdbase_exist = TRUE )
+#   # ONLY LINK to a permanent GRASS folder at "root_folder", location named "project1"
+#   linkGRASS(gisdbase = root_folder, location = "project1", gisdbase_exist = TRUE )
 #  
 #  
 #   # setting up GRASS manually with spatial parameters of the nc data
@@ -91,9 +89,9 @@
 #   linkGRASS(spatial_params = c(178605,329714,181390,333611,proj4_string))
 #  
 #   # creating a GRASS gisdbase manually with spatial parameters of the nc data
-#   # additionally using a peramanent directory "projRootDir" and the location "nc_spatial_params "
+#   # additionally using a peramanent directory "root_folder" and the location "nc_spatial_params "
 #   proj4_string <- as.character(sp::CRS("+init=epsg:4267"))
-#   linkGRASS(gisdbase = projRootDir,
+#   linkGRASS(gisdbase = root_folder,
 #              location = "nc_spatial_params",
 #              spatial_params = c(-84.32385, 33.88199,-75.45698,36.58965,proj4_string))
 #  
@@ -159,7 +157,7 @@
 #  require(listviewer)
 #  
 #  otblink<-link2GI::linkOTB()
-#   projRootDir<-tempdir()
+#   root_folder<-tempdir()
 #  
 #  fn <- system.file("ex/elev.tif", package = "terra")
 #  
@@ -176,7 +174,7 @@
 #  cmd$input  <- fn
 #  cmd$filter <- "touzi"
 #  cmd$channel <- 2
-#  cmd$out <- file.path(projRootDir,paste0("out",cmd$filter,".tif"))
+#  cmd$out <- file.path(root_folder,paste0("out",cmd$filter,".tif"))
 #  
 #  ## run algorithm
 #  retStack<-runOTB(cmd,gili = otblink)