+++ vignettes/link2GI3.R
@@ -1,16 +1,15 @@
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #   # we need some additional packages
 #   require(link2GI)
 #   require(curl)
 #  
 #  # first of all we create  a project folder structure
-#    link2GI::initProj(projRootDir = paste0(tempdir(),"/link2GI_examples"),
-#                      projFolders =  c("run/"),
-#                      path_prefix = "path_",
-#                      global = TRUE)
+#    envrmt = link2GI::createFolders(root_folder = paste0(tempdir(),"/link2GI_examples"),
+#                                    folders =  c("run/"),
+#                                    path_prefix = "path_")
 #  
 #  # set runtime directory
-#    setwd(path_run)
+#    setwd(envrmt$path_run)
 #  
 #  # get some typical authority generated data
 #    url<-"https://www.zensus2011.de/SharedDocs/Downloads/DE/Pressemitteilung/DemografischeGrunddaten/csv_Bevoelkerung_100m_Gitter.zip;jsessionid=294313DDBB57914D6636DE373897A3F2.2_cid389?__blob=publicationFile&v=3"
@@ -22,13 +21,13 @@
 #  fn <- list.files(pattern = "[.]csv$", path = getwd(), full.names = TRUE)
 #  
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  # fast read with data.table
 #   xyz <- data.table::fread(paste0(path_run,"/Zensus_Bevoelkerung_100m-Gitter.csv"))
 #  
 #   head(xyz)
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #   require(RColorBrewer)
 #   require(terra)
 #   require(mapview)
@@ -50,14 +49,14 @@
 #                    legend = TRUE)
 #  
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  require(link2GI)
 #  # initialize GRASS and set up a permanent structure
 #  link2GI::linkGRASS(x = r,
 #                      gisdbase = paste0(tempdir(),"/link2GI_examples"),
 #                      location = "microzensus2011")
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  require(link2GI)
 #  require(raster)
 #  require(rgrass)
@@ -77,7 +76,7 @@
 #  rgrass::execGRASS('r.info',
 #                     map = "Zensus_Bevoelkerung_100m_Gitter")
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  
 #   xyz_sf = st_as_sf(xyz,
 #                      coords = c("x_mp_100m", "y_mp_100m"),
@@ -87,7 +86,7 @@
 #  #map points
 #   sf::plot_sf(xyz_sf)
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #   require(sf)
 #   require(sp)
 #   require(link2GI)